refactored the Tree conversion code into its own class to simplify ConceptGraph and ConceptNode and tests
