+++ test/edu/ithaca/dragonlab/ckc/conceptgraph/ConceptGraphTest.java
@@ -234,79 +234,6 @@ public class ConceptGraphTest {
         Assert.assertEquals(0.5,   graph.findNodeById("C").getLearningObjectMap().get("Q6").calcKnowledgeEstimate(), DataUtil.OK_FLOAT_MARGIN);
     }
 
-    @Test
-    public void simpleCheckNumEachIDTest(){
-        //checks if the tree has the right number of copies per orig node
-        int numA = 0;
-        int numB = 0;
-        int numC = 0;
-
-        ConceptGraphRecord treeLists = ExampleConceptGraphFactory.makeSimpleWithData().graphToTree().buildNodesAndLinks();
-        for(ConceptRecord node : treeLists.getConcepts()){
-			if(node.getLabel().equals("A")){
-                numA++;
-            }else if(node.getLabel().equals("B")){
-                numB++;
-            }else if(node.getLabel().equals("C")){
-                numC++;
-            }
-        }
-        Assert.assertEquals(1, numA);
-        Assert.assertEquals(1, numB);
-        Assert.assertEquals(2, numC);
-    }
-
-    @Test
-	public void treeConversionTest(){
-		checkTreeConversionByNodesAndLinksNumbers(ExampleConceptGraphFactory.makeSimpleWithData(), 3, 3, 4, 3);
-		checkTreeConversionByNodesAndLinksNumbers(ExampleConceptGraphFactory.makeMedium(), 4, 5, 7, 6);
-		checkTreeConversionByNodesAndLinksNumbers(ExampleConceptGraphFactory.makeComplex(), 5, 8, 13, 12);
-		checkTreeConversionByNodesAndLinksNumbers(ExampleConceptGraphFactory.makeSuperComplex(), 6, 11, 24, 23);
-	}
-
-	public void checkTreeConversionByNodesAndLinksNumbers(ConceptGraph graphToTest, int expectedGraphNodeCount, int expectedGraphLinkCount, int expectedTreeNodeCount, int expectedTreeLinkCount){
-		ConceptGraphRecord graphLists = graphToTest.buildNodesAndLinks();
-		Assert.assertEquals(expectedGraphNodeCount, graphLists.getConcepts().size());
-		Assert.assertEquals(expectedGraphLinkCount, graphLists.getLinks().size());
-
-		ConceptGraphRecord treeLists = graphToTest.graphToTree().buildNodesAndLinks();
-
-		Assert.assertEquals(expectedTreeNodeCount, treeLists.getConcepts().size());
-		Assert.assertEquals(expectedTreeLinkCount, treeLists.getLinks().size());
-	}
-
-	//TODO: fis the original creators
-    @Test
-    public void treeToTreeTest(){
-//        treeToTreeCheck(ExampleConceptGraphFactory.makeSimpleInputTree());
-//        treeToTreeCheck(ExampleConceptGraphFactory.makeComplexInputTree());
-    }
-
-    public void treeToTreeCheck(ConceptGraph treeToTest){
-        ConceptGraphRecord initialLists = treeToTest.buildNodesAndLinks();
-        ConceptGraph treeFromTree = treeToTest.graphToTree();
-
-        ConceptGraphRecord postLists = treeFromTree.buildNodesAndLinks();
-        Assert.assertEquals(postLists.getConcepts().size(), initialLists.getConcepts().size());
-        Assert.assertEquals(postLists.getLinks().size(), initialLists.getLinks().size());
-    }
-
-    @Test
-    public void mediumReTreeTest(){
-        ConceptGraph myGraph = ExampleConceptGraphFactory.makeMedium().graphToTree();
-        ConceptGraph myTree = myGraph.graphToTree();
-        ConceptGraph myTree2 = myTree.graphToTree();
-
-        ConceptGraphRecord lists1 = myGraph.buildNodesAndLinks();
-        ConceptGraphRecord lists2 = myTree.buildNodesAndLinks();
-        ConceptGraphRecord lists3 = myTree2.buildNodesAndLinks();
-
-        Assert.assertEquals(lists2.getConcepts().size(), lists3.getConcepts().size());
-
-        Assert.assertEquals(lists2.getLinks().size(), lists3.getLinks().size());
-
-        Assert.assertEquals(lists3.getLinks().size() + 1, lists3.getConcepts().size());
-    }
 
     //TODO: Don't know how these tests are still passing, I think they should be broken, so they should be re-written to catch errors, then the format of test files should be fixed
 	@Test
