+++ test/edu/ithaca/dragonlab/ckc/conceptgraph/PracticalConceptGraphTest.java
@@ -30,7 +30,7 @@ public class PracticalConceptGraphTest {
             ConceptGraph graph = new ConceptGraph(graphRecord, LOLRlist, csvReader.getManualGradedResponses());
 
             ObjectMapper mapper = new ObjectMapper();
-            mapper.writeValue(new File("test/testresources/practicalExamples/blankRealisticExample.json"), graph.graphToTree().buildNodesAndLinks());
+            mapper.writeValue(new File("test/testresources/practicalExamples/blankRealisticExample.json"), TreeConverter.makeTreeCopy(graph).buildNodesAndLinks());
 
             graph.calcDataImportance();
             graph.calcKnowledgeEstimates();
@@ -39,7 +39,7 @@ public class PracticalConceptGraphTest {
             Assert.assertEquals(5,graph.findNodeById("Boolean").getLearningObjectMap().size());
             Assert.assertEquals(15,graph.getLearningObjectMap().size());
 
-            ConceptGraphRecord tree = graph.graphToTree().buildNodesAndLinks();
+            ConceptGraphRecord tree = TreeConverter.makeTreeCopy(graph).buildNodesAndLinks();
             //Object to JSON in file
             mapper = new ObjectMapper();
             mapper.writeValue(new File("test/testresources/practicalExamples/basicRealisticExample.json"), tree);
@@ -73,7 +73,7 @@ public class PracticalConceptGraphTest {
             Assert.assertEquals(15,graph.getLearningObjectMap().size());
 
             ObjectMapper mapper = new ObjectMapper();
-            ConceptGraphRecord tree = graph.graphToTree().buildNodesAndLinks();
+            ConceptGraphRecord tree = TreeConverter.makeTreeCopy(graph).buildNodesAndLinks();
             //Object to JSON in file
             mapper.writeValue(new File("test/testresources/practicalExamples/advancedRealisticExample.json"), tree);
             //mapper.writeValue(new File("test/testresources/practicalExamples/groupConceptGraphAdvancedRealisticExample.json"), gcg.getAllNamedGraphs());
@@ -114,7 +114,7 @@ public class PracticalConceptGraphTest {
 
 
 
-            ConceptGraphRecord tree = graph.graphToTree().buildNodesAndLinks();
+            ConceptGraphRecord tree = TreeConverter.makeTreeCopy(graph).buildNodesAndLinks();
             ObjectMapper mapper = new ObjectMapper();
             //Object to JSON in file
             mapper.writeValue(new File("test/testresources/practicalExamples/singleStudentRealisticExample.json"), tree);
